pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "code-scan-agent"
version = "0.1.0"
description = "Agent scan code sử dụng ADK và Semgrep MCP"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Code Scan Agent", email = "contact@example.com" },
]
dependencies = [
    "google-adk>=1.3.0",
    "python-dotenv>=1.0.0",
    "typing-extensions",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "black>=24.0",
    "isort>=5.13",
    "mypy>=1.8",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 
